requirements.txt
APScheduler == 3.9.1
aiofiles == 0.8.0
aiohttp == 3.8.1
arrow == 1.2.2
bs4 == 0.0.1
discord.py == 1.7.3
ics == 0.7
lxml == 4.8.0
requests == 2.27.1
